+++ R/Fox.R
@@ -13,7 +13,6 @@
 #' @param gen The name of the column that contains the levels of the genotypes.
 #' @param resp The response variable(s). To analyze multiple variables in a
 #'   single procedure use, for example, \code{resp = c(var1, var2, var3)}.
-#' @param rep \strong{Deprecated argument. It will be retired in the next release.}
 #' @param verbose Logical argument. If \code{verbose = FALSE} the code will run
 #'   silently.
 #' @return An object of class \code{Fox}, which is a list containing the results
@@ -37,10 +36,7 @@
 #' print(out)
 #'}
 #'
-Fox <- function(.data, env, gen, resp, rep = "deprecated", verbose = TRUE) {
-  if(rep != "deprecated"){
-    warning("`verbose` is deprecated. It will be defunct in the new release.", call. = FALSE)
-  }
+Fox <- function(.data, env, gen, resp, verbose = TRUE) {
   factors  <-
     .data %>%
     select({{env}}, {{gen}}) %>%