+++ R/utilities.R
@@ -1023,31 +1023,6 @@ hmean <- function(.data, ..., na.rm = FALSE) {
 }
 #' @name utils_stats
 #' @export
-hm_mean <- function(.data, ..., na.rm = FALSE) {
-  .Deprecated(msg = "`hm_mean()` is deprecated. Use `hmean()` instead.")
-  funct <- function(df){
-    1 / mean(1 / df, na.rm = na.rm)
-  }
-  if(has_na(.data) && na.rm == FALSE){
-    stop("NA values in data. Use 'na.rm = TRUE' to remove NAs from analysis.\nTo remove rows with NA use `remove_rows_na()'. \nTo remove columns with NA use `remove_cols_na()'.", call. = FALSE)
-  }
-  if(is.null(nrow(.data))){
-    funct(.data)
-  } else{
-    if(missing(...)){
-      .data %>%
-        summarise_if(is.numeric, funct) %>%
-        ungroup()
-    } else{
-      .data %>%
-       select_cols(group_vars(.), ...) %>%
-        summarise_if(is.numeric, funct) %>%
-        ungroup()
-    }
-  }
-}
-#' @name utils_stats
-#' @export
 gmean <- function(.data, ..., na.rm = FALSE){
   funct <- function(df){
     exp(sum(log(df[df > 0]), na.rm = na.rm) / length(df))
@@ -1065,31 +1040,6 @@ gmean <- function(.data, ..., na.rm = FALSE){
     } else{
       .data %>%
         select_cols(group_vars(.), ...) %>%
-        summarise_if(is.numeric, funct) %>%
-        ungroup()
-    }
-  }
-}
-#' @name utils_stats
-#' @export
-gm_mean <- function(.data, ..., na.rm = FALSE){
-  .Deprecated(msg = "`gm_mean()` is deprecated. Use `gmean()` instead.")
-  funct <- function(df){
-    exp(sum(log(df[df > 0]), na.rm = na.rm) / length(df))
-  }
-  if(has_na(.data) && na.rm == FALSE){
-    stop("NA values in data. Use 'na.rm = TRUE' to remove NAs from analysis.\nTo remove rows with NA use `remove_rows_na()'. \nTo remove columns with NA use `remove_cols_na()'.", call. = FALSE)
-  }
-  if(is.null(nrow(.data))){
-    funct(.data)
-  } else{
-    if(missing(...)){
-      .data %>%
-        summarise_if(is.numeric, funct) %>%
-        ungroup()
-    } else{
-      .data %>%
-       select_cols(group_vars(.), ...) %>%
         summarise_if(is.numeric, funct) %>%
         ungroup()
     }